Use array in "write to files" test.
The use of sizeof to get the size of a string pointed to by a char * is probably not what was intended. 
The test will report success, but the entire string will not be written. 
Need to either make str into a static array so that sizeof does the right thing, or use strlen to calculate its size.
Used the former in this commit.